# bazantic/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lute Audit API",
    "version": "0.1.0",
    "description": "Independent verification of a Subgraph index against raw Base RPC logs. POST an ERC-4626 vault + block range and get a machine-readable reconciliation verdict (VERIFIED / FAILED / INCONCLUSIVE) with first-divergence evidence. Bazantic fronts this upstream with x402/MPP payments and exposes each operation as an MCP recipe."
  },
  "servers": [
    { "url": "https://uselute.xyz", "description": "Public Lute verification API" }
  ],
  "paths": {
    "/v1/audits": {
      "post": {
        "operationId": "auditVaultV1",
        "summary": "Run an independent ERC-4626 audit",
        "description": "Runs the Lute verifier against raw Base RPC logs and an indexed candidate.",
        "parameters": [
          { "$ref": "#/components/parameters/ContractQuery" },
          { "$ref": "#/components/parameters/EventQuery" },
          { "$ref": "#/components/parameters/FromBlockQuery" },
          { "$ref": "#/components/parameters/ToBlockQuery" },
          { "$ref": "#/components/parameters/SubgraphQuery" }
        ],
        "responses": {
          "200": { "description": "Audit report", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuditReport" } } } },
          "400": { "description": "Invalid request" },
          "500": { "description": "Audit infrastructure failure" }
        }
      }
    },
    "/v1/verifications": {
      "post": {
        "operationId": "verifyCurrentCandidate",
        "summary": "Verify and persist the configured candidate",
        "description": "Runs the audit, hashes the server-configured candidate, and persists an immutable-ish VerificationRun used by the deployment gate.",
        "parameters": [
          { "$ref": "#/components/parameters/ContractQuery" },
          { "$ref": "#/components/parameters/EventQuery" },
          { "$ref": "#/components/parameters/FromBlockQuery" },
          { "$ref": "#/components/parameters/ToBlockQuery" },
          { "$ref": "#/components/parameters/SubgraphQuery" }
        ],
        "responses": {
          "201": { "description": "Verified run persisted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VerificationRun" } } } },
          "200": { "description": "Run persisted with FAILED or INCONCLUSIVE verdict" }
        }
      }
    },
    "/v1/verifications/{runId}": {
      "get": {
        "operationId": "getVerification",
        "summary": "Get a persisted VerificationRun",
        "parameters": [{ "$ref": "#/components/parameters/RunId" }],
        "responses": { "200": { "description": "VerificationRun", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VerificationRun" } } } }, "404": { "description": "Run not found" } }
      }
    },
    "/v1/verifications/{runId}/evidence": {
      "get": {
        "operationId": "getVerificationEvidence",
        "summary": "Get evidence for a persisted run",
        "parameters": [{ "$ref": "#/components/parameters/RunId" }],
        "responses": { "200": { "description": "Evidence and root" } }
      }
    },
    "/v1/verifications/{runId}/manifest": {
      "get": {
        "operationId": "getVerifiedManifest",
        "summary": "Get the candidate manifest bound to a run",
        "parameters": [{ "$ref": "#/components/parameters/RunId" }],
        "responses": { "200": { "description": "Candidate manifest" } }
      }
    },
    "/v1/verifications/{runId}/pack": {
      "get": {
        "operationId": "getIntegrityPackArtifact",
        "summary": "Get the portable Integrity Pack artifact for a run",
        "description": "Returns the reviewed pack definition, TrustManifest, candidate hash, complete audit evidence, stable lineage ids, and public payment receipt metadata when the run was paid.",
        "parameters": [{ "$ref": "#/components/parameters/RunId" }],
        "responses": { "200": { "description": "Portable Integrity Pack artifact", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IntegrityPackArtifact" } } } }, "404": { "description": "Run not found" } }
      }
    },
    "/v1/integrity-packs": {
      "get": { "operationId": "listIntegrityPacks", "summary": "List supported Integrity Packs", "responses": { "200": { "description": "Integrity Packs" } } }
    },
    "/v1/integrity-packs/erc4626@1": {
      "get": { "operationId": "getErc4626Pack", "summary": "Get the ERC-4626 Integrity Pack", "responses": { "200": { "description": "ERC-4626 pack" } } }
    },
    "/api/audit": {
      "post": {
        "operationId": "auditVault",
        "summary": "Reconcile a Subgraph index against raw chain data",
        "description": "Reconstructs the event set directly from Base RPC logs and compares it to the same events as reported by an index. Returns the verdict and, on disagreement, the first divergent (block, tx, logIndex).",
        "parameters": [
          { "$ref": "#/components/parameters/ContractQuery" },
          { "$ref": "#/components/parameters/EventQuery" },
          { "$ref": "#/components/parameters/FromBlockQuery" },
          { "$ref": "#/components/parameters/ToBlockQuery" },
          { "$ref": "#/components/parameters/SubgraphQuery" }
        ],
        "responses": {
          "200": {
            "description": "Audit report",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuditReport" } } }
          },
          "400": { "description": "Invalid request" }
        }
      }
    },
    "/api/events": {
      "get": {
        "operationId": "listSupportedEvents",
        "summary": "List supported ERC-4626 events",
        "description": "Returns the events Lute can verify with their canonical signature and keccak256-derived topic0.",
        "responses": {
          "200": {
            "description": "Supported events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": { "type": "string" },
                      "signature": { "type": "string" },
                      "topic0": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuditRequest": {
        "type": "object",
        "required": ["contract", "fromBlock", "toBlock"],
        "properties": {
          "contract": { "type": "string", "pattern": "^0x[0-9a-fA-F]{40}$", "description": "ERC-4626 vault address on Base" },
          "event": { "type": "string", "enum": ["Deposit", "Withdraw"], "default": "Deposit" },
          "fromBlock": { "oneOf": [{ "type": "integer", "minimum": 0 }, { "type": "string", "pattern": "^[0-9]+$" }], "description": "inclusive start block; strings preserve large integers" },
          "toBlock": { "oneOf": [{ "type": "integer", "minimum": 0 }, { "type": "string", "pattern": "^[0-9]+$" }], "description": "inclusive end block (well behind chain head)" },
          "subgraph": { "type": "string", "default": "morpho", "description": "morpho | graphstudio | graphnode:<name> | local[:bug]" }
        }
      },
      "AuditReport": {
        "type": "object",
        "properties": {
          "runId": { "type": "string" },
          "event": { "type": "string" },
          "verdict": { "type": "string", "enum": ["VERIFIED", "FAILED", "INCONCLUSIVE"] },
          "eventsChecked": { "type": "integer" },
          "range": {
            "type": "object",
            "properties": {
              "startBlock": { "type": "string" },
              "endBlock": { "type": "string" },
              "safeHead": { "type": "string" }
            }
          },
          "checks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "status": { "type": "string", "enum": ["PASS", "FAIL", "UNVERIFIED"] },
                "detail": { "type": "string" }
              }
            }
          },
          "rawEvidence": { "type": "object", "description": "RAW_RPC provenance (topic0, chunkCount, eventCount, rpcAlias)" },
          "subgraphEvidence": { "type": "object", "description": "index provenance (endpoint, entity, pageCount, recordCount)" },
          "evidenceRoot": { "type": "string", "pattern": "^[0-9a-f]{64}$" },
          "firstDivergence": {
            "type": ["object", "null"],
            "properties": {
              "blockNumber": { "type": "string" },
              "transactionHash": { "type": "string" },
              "logIndex": { "type": "integer" },
              "event": { "type": "string" },
              "check": { "type": "string" }
            }
          }
        }
      },
      "VerificationRun": {
        "type": "object",
        "required": ["schemaVersion", "runId", "candidateHash", "report", "coverage", "evidenceRoot", "verdict"],
        "properties": {
          "schemaVersion": { "const": 1 },
          "runId": { "type": "string" },
          "candidateHash": { "type": "string", "pattern": "^[0-9a-f]{64}$" },
          "candidate": { "type": "object" },
          "report": { "$ref": "#/components/schemas/AuditReport" },
          "coverage": { "type": "object" },
          "evidenceRoot": { "type": "string", "pattern": "^[0-9a-f]{64}$" },
          "verdict": { "type": "string", "enum": ["VERIFIED", "FAILED", "INCONCLUSIVE"] },
          "revoked": { "type": "boolean" }
        }
      },
      "IntegrityPackArtifact": {
        "type": "object",
        "required": ["schemaVersion", "pack", "trustManifest", "verification", "candidate", "evidence"],
        "properties": {
          "schemaVersion": { "const": 1 },
          "pack": { "type": "object", "description": "The reviewed pack.yaml definition" },
          "trustManifest": { "type": "object", "description": "Stable trust binding for verdict, candidate, subject, sources, and evidence root" },
          "verification": { "type": "object", "description": "Complete AuditReport and coverage" },
          "candidate": { "type": "object", "description": "Candidate hash and deterministic file manifest" },
          "evidence": { "type": "object", "description": "Evidence root and stable block/source/check lineage ids" },
          "payment": { "type": "object", "description": "Public x402 settlement metadata; never contains signing material" }
        }
      }
    },
    "parameters": {
      "ContractQuery": {
        "name": "contract",
        "in": "query",
        "required": true,
        "schema": { "type": "string", "pattern": "^0x[0-9a-fA-F]{40}$" },
        "description": "ERC-4626 vault address on Base"
      },
      "EventQuery": {
        "name": "event",
        "in": "query",
        "required": false,
        "schema": { "type": "string", "enum": ["Deposit", "Withdraw"], "default": "Deposit" }
      },
      "FromBlockQuery": {
        "name": "fromBlock",
        "in": "query",
        "required": true,
        "schema": { "type": "string", "pattern": "^[0-9]+$" },
        "description": "inclusive start block; decimal string"
      },
      "ToBlockQuery": {
        "name": "toBlock",
        "in": "query",
        "required": true,
        "schema": { "type": "string", "pattern": "^[0-9]+$" },
        "description": "inclusive end block; decimal string"
      },
      "SubgraphQuery": {
        "name": "subgraph",
        "in": "query",
        "required": false,
        "schema": { "type": "string", "default": "morpho" },
        "description": "morpho | graphstudio | graphnode:<name> | local[:bug]"
      },
      "RunId": {
        "name": "runId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "pattern": "^[A-Za-z0-9_-]+$" }
      }
    }
  }
}